# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ICreateCategoryDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUpdateCategoryDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"featured": {
						"type": "boolean"
					},
					"menu": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBrandCreateDTO": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBrandUpdateDTO": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IProductCreateDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"countInStock": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"type": "string"
					},
					"brand": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"description",
					"price",
					"countInStock",
					"category",
					"brand"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAttributeCreateDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"frontend_type": {
						"type": "string"
					},
					"is_filterable": {
						"type": "boolean"
					},
					"is_required": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"code",
					"frontend_type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAttributeUpdateDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"frontend_type": {
						"type": "string"
					},
					"is_filterable": {
						"type": "boolean"
					},
					"is_required": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"code",
					"frontend_type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAttributeValueCreateDTO": {
				"properties": {
					"value": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"attribute": {
						"type": "string"
					}
				},
				"required": [
					"value",
					"attribute"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAttributeValueUpdateDTO": {
				"properties": {
					"value": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"attribute": {
						"type": "string"
					}
				},
				"required": [
					"value",
					"attribute"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"description": "JWT Authorization header using the Bearer scheme.\n\nEnter your token in the text input below.\n\n Example: \"12345abcdef\"\n\n Name: Authorization\n\nIn: header"
			}
		}
	},
	"info": {
		"title": "master-project",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/admin/category/create": {
			"post": {
				"operationId": "CreateCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Create new category",
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateCategoryDTO"
							}
						}
					}
				}
			}
		},
		"/admin/category/all": {
			"get": {
				"operationId": "GetAllCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Return all categories from DB",
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/admin/category/{categoryId}": {
			"get": {
				"operationId": "GetCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Category details",
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/category/update/{categoryId}": {
			"put": {
				"operationId": "UpdateCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Update category",
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUpdateCategoryDTO"
							}
						}
					}
				}
			}
		},
		"/admin/category/delete/{categoryId}": {
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"description": "Delete category",
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/category/image-upload/{categoryId}": {
			"put": {
				"operationId": "UploadCategoryImage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Upload category image",
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"image": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"image"
								]
							}
						}
					}
				}
			}
		},
		"/admin/brand/create": {
			"post": {
				"operationId": "CreateBrand",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Create new brand",
				"tags": [
					"Brands"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IBrandCreateDTO"
							}
						}
					}
				}
			}
		},
		"/admin/brand/all": {
			"get": {
				"operationId": "GetAllBrands",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Return all brands from DB",
				"tags": [
					"Brands"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/admin/brand/{brandId}": {
			"get": {
				"operationId": "GetBrand",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Brand details",
				"tags": [
					"Brands"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "brandId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/brand/update/{brandId}": {
			"put": {
				"operationId": "UpdateBrand",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Update brand",
				"tags": [
					"Brands"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "brandId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IBrandUpdateDTO"
							}
						}
					}
				}
			}
		},
		"/admin/brand/delete/{brandId}": {
			"delete": {
				"operationId": "DeleteBrand",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"description": "Delete brand",
				"tags": [
					"Brands"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "brandId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/brand/image-upload/{brandId}": {
			"put": {
				"operationId": "UploadBrandImage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Upload brand image",
				"tags": [
					"Brands"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "brandId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"logo": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"logo"
								]
							}
						}
					}
				}
			}
		},
		"/admin/product/create": {
			"post": {
				"operationId": "CreateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProductCreateDTO"
							}
						}
					}
				}
			}
		},
		"/admin/product/all": {
			"get": {
				"operationId": "GetAllProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/admin/product/{productId}": {
			"get": {
				"operationId": "GetProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/attribute/create": {
			"post": {
				"operationId": "CreateAttribute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Create new attribute",
				"tags": [
					"Attributes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAttributeCreateDTO"
							}
						}
					}
				}
			}
		},
		"/admin/attribute/all": {
			"get": {
				"operationId": "GetAllAttributes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Retrieve all attributes from DB",
				"tags": [
					"Attributes"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/admin/attribute/{attributeId}": {
			"get": {
				"operationId": "GetAttribute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Attribute details",
				"tags": [
					"Attributes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "attributeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/attribute/update/{attributeId}": {
			"put": {
				"operationId": "UpdateAttribute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Update attribute",
				"tags": [
					"Attributes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "attributeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAttributeUpdateDTO"
							}
						}
					}
				}
			}
		},
		"/admin/attribute/delete/{attributeId}": {
			"delete": {
				"operationId": "DeleteAttribute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"description": "Delete attribute",
				"tags": [
					"Attributes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "attributeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/attribute-value/create": {
			"post": {
				"operationId": "CreateAttributeValue",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Create a new attribute value",
				"tags": [
					"Attribute Values"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAttributeValueCreateDTO"
							}
						}
					}
				}
			}
		},
		"/admin/attribute-value/all": {
			"get": {
				"operationId": "GetAllAttributeValues",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Retrieve all attribute value from DB",
				"tags": [
					"Attribute Values"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/admin/attribute-value/{attributeValueId}": {
			"get": {
				"operationId": "GetAttributeValue",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Attribute value details",
				"tags": [
					"Attribute Values"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "attributeValueId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/attribute-value/update/{attributeValueId}": {
			"put": {
				"operationId": "UpdateAttributeValue",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Update attribute value",
				"tags": [
					"Attribute Values"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "attributeValueId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAttributeValueUpdateDTO"
							}
						}
					}
				}
			}
		},
		"/admin/attribute-value/delete/{attributeValueId}": {
			"delete": {
				"operationId": "DeleteAttributeValue",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Delete attribute value",
				"tags": [
					"Attribute Values"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "attributeValueId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}